shipit_signoff: enable back tests
To be able to enable tests we needed to move flask application which was
created in ``__init__.py`` to ``flask.py``. That allowed us to import
things from our project.

This change was also applied to other applications.
