+++ src/shipit_signoff/tests/test_api.py
@@ -5,17 +5,14 @@ import yaml
 
 
 HERE = path.abspath(path.dirname(__file__))
+SIGNOFF_DIR = path.join(HERE, '..', 'shipit_signoff')
 
 
 def test_swagger_spec():
-    import shipit_signoff
-
-    shipit_dir = path.abspath(path.dirname(shipit_signoff.__file__))
-
     with open(path.join(HERE, "swagger-2.0-spec.json")) as f:
         swagger_schema = json.load(f)
 
-    with open(path.join(shipit_dir, "api.yml")) as f:
+    with open(path.join(SIGNOFF_DIR, "api.yml")) as f:
         swagger_file = yaml.load(f.read())
 
     jsonschema.validate(swagger_file, swagger_schema)
